requirements.docker.txt
lightning==2.6.0
torchmetrics==1.8.2
tensorboard==2.20.0

